refactor: use built in fractions module
